Progress on Bubble Sort - FIXME ArgumentError
It is close, but breaks because of some indexing. The for loop inside the until loop needs to be fixed becasue it does not iterate through the array properly.
After rewatching the Harvard CS50 lesson again, I realized that my
earlier approach was for a different algorithm (was not committed here
as I was working in replit). I need to find the
largest number in the array, sort that to the end of the array, solidify
that position, then find the next largest number in array (0..n-1) then
repeat (0..n-2). At first, I was trying to follow the video and use k as well,
but I had difficulty - I am not sure if I need to incorporate with how I
have things set up.