Finish with bubble sort
It took a long while to rework the is_sorted? method. I don't think it
was necessary to make a copy - I kept it from previous versions of
is_sorted?. Instead of removing values from the array, it has a count.
I use the add variable in the if/else statement because I don't want to
permenantly set the count value. I want to add 1 or 0 when we check each
number against its neighbor. I also added the if (index+1) ==
copy.length because we want to break out of the loop once we finish
evaluating all the numbers. For instance, in the example, we don't have
any number after 78 and want to ensure we break.